+++ networkml/featurizers/funcs/host.py
@@ -11,7 +11,7 @@ ETH_TYPE_IPX = 0x8137
 ETH_IP_TYPES = frozenset((ETH_TYPE_ARP, ETH_TYPE_IP, ETH_TYPE_IPV6))
 
 
-class Host(Features):
+class HostBase:
 
     NAME_TO_STAT = {
         'count': len,
@@ -36,41 +36,26 @@ class Host(Features):
     WK_PROTOS = frozenset(('tcp', 'udp', 'icmp', 'icmpv6', 'arp', 'other'))
 
 
-    @staticmethod
-    def _get_ips(row):
-        ip_src = None
-        for src_field in ('ip.src', 'ip.src_host'):
-            ip_src = row.get(src_field, None)
-            if ip_src:
-                break
-        ip_dst = None
-        for dst_field in ('ip.dst', 'ip.dst_host'):
-            ip_dst = row.get(dst_field, None)
-            if ip_dst:
-                break
-        if ip_src and ip_dst:
-            ip_src = ipaddress.ip_address(ip_src)
-            ip_dst = ipaddress.ip_address(ip_dst)
-        return (ip_src, ip_dst)
-
+    def _pyshark_ipversions(self, rows):
+        ipversions = set()
+        for row in self._pyshark_row_layers(rows):
+            if '<IP Layer>' in row['layers']:
+                ipversions.add(4)
+            elif '<IPV6 Layer>' in row['layers']:
+                ipversions.add(6)
+        return ipversions
 
     @staticmethod
-    def _get_proto_eth_type(row):
-        for eth_field in ('vlan.etype', 'eth.type'):
-            eth_type = row.get(eth_field, None)
-            if eth_type:
-                return eth_type
-        return 0
+    def _last_protocols(rows):
+        protocols = ''
+        for row in rows:
+            row_protocols = row.get('frame.protocols', None)
+            if row_protocols is not None:
+                protocols = row_protocols
+        return protocols
 
 
-    @staticmethod
-    def _safe_int(maybeint):
-        if isinstance(maybeint, int) or maybeint is None:
-            return maybeint
-        if maybeint:
-            return int(maybeint, 0)
-        return None
-
+class Host(HostBase, Features):
 
     def _row_keys(self, row):
         keys = set()
@@ -83,14 +68,12 @@ class Host(Features):
             keys.add(val)
         return keys
 
-
     def _partition_rows_by_field_vals(self, rows):
         partitioned_rows = defaultdict(list)
         for row in rows:
             for key in self._row_keys(row):
                 partitioned_rows[key].append(row)
         return partitioned_rows
-
 
     def _host_rows(self, rows, host_func):
         newrows = []
@@ -100,24 +83,11 @@ class Host(Features):
             newrows.append(host_func_results)
         return newrows
 
-
-    def _pyshark_ipversions(self, rows):
-        ipversions = set()
-        for row in self._pyshark_row_layers(rows):
-            if '<IP Layer>' in row['layers']:
-                ipversions.add(4)
-            elif '<IPV6 Layer>' in row['layers']:
-                ipversions.add(6)
-        return ipversions
-
-
     def pyshark_ipv4(self, rows):
         return [{'IPv4': (4 in self._pyshark_ipversions(rows))}]
 
-
     def pyshark_ipv6(self, rows):
         return [{'IPv6': (6 in self._pyshark_ipversions(rows))}]
-
 
     def pyshark_last_highest_layer(self, rows):
         highest_layer = 0
@@ -125,49 +95,12 @@ class Host(Features):
             highest_layer = row['layers'].split('<')[-1]
         return [{'highest_layer': highest_layer}]
 
-
     def pyshark_layers(self, rows):
         layers = set()
         for row in self._pyshark_row_layers(rows):
             temp = row['layers'].split('<')[1:]
             layers.update({layer.split(' Layer')[0] for layer in temp})
         return [{layer: 1 for layer in layers}]
-
-
-    @staticmethod
-    def _last_protocols(rows):
-        protocols = ''
-        for row in rows:
-            row_protocols = row.get('frame.protocols', None)
-            if row_protocols is not None:
-                protocols = row_protocols
-        return protocols
-
-
-    def tshark_last_protocols_array(self, rows):
-
-        def last_protocols_array(host_rows):
-            raw_protocols = set()
-            try:
-                raw_protocols.update({
-                    protocol for protocol in self._last_protocols(host_rows).split(':') if protocol})
-            except IndexError:
-                pass
-            raw_protocols -= set(['ethertype'])
-            protocols = {'protocol_%s' % protocol: int(protocol in raw_protocols) for protocol in self.WK_PROTOCOLS}
-            protocols.update({'other': int(not raw_protocols.issubset(self.WK_PROTOCOLS))})
-            return protocols
-
-        return self._host_rows(rows, last_protocols_array)
-
-
-    def tshark_ipv4(self, rows):
-        return self._host_rows(rows, lambda x: {'IPv4': int(4 in self._tshark_ipversions(x))})
-
-
-    def tshark_ipv6(self, rows):
-        return self._host_rows(rows, lambda x: {'IPv6': int(6 in self._tshark_ipversions(x))})
-
 
     def _calc_tshark_field(self, field, tshark_field, rows):
 
@@ -186,22 +119,18 @@ class Host(Features):
 
         return self._host_rows(rows, calc_field)
 
-
     def _calc_time_delta(self, field, rows):
         assert 'time_delta' in field
         return self._calc_tshark_field(field, 'frame.time_delta_displayed', rows)
-
 
     def _calc_framelen(self, field, rows):
         assert 'frame_len' in field
         return self._calc_tshark_field(field, 'frame.len', rows)
 
-
     def _get_ip_proto_ports(self, row, ip_proto):
         src_port = self._safe_int(row.get('.'.join((ip_proto, 'srcport')), None))
         dst_port = self._safe_int(row.get('.'.join((ip_proto, 'dstport')), None))
         return (src_port, dst_port)
-
 
     def _lowest_ip_proto_ports(self, rows, ip_proto):
         lowest_ports = set()
@@ -212,20 +141,17 @@ class Host(Features):
                 lowest_ports.add(min_port)
         return lowest_ports
 
-
     def _priv_ip_proto_ports(self, rows, ip_proto):
         lowest_ports = {port for port in self._lowest_ip_proto_ports(rows, ip_proto) if port < 1024}
         priv_ports = {port: int(port in lowest_ports) for port in self.WK_PRIV_PROTO_PORTS}
         priv_ports.update({'other': int(not lowest_ports.issubset(self.WK_PRIV_PROTO_PORTS))})
         return priv_ports
 
-
     def _nonpriv_ip_proto_ports(self, rows, ip_proto):
         lowest_ports = {port for port in self._lowest_ip_proto_ports(rows, ip_proto) if port >= 1024}
         nonpriv_ports = {port: int(port in lowest_ports) for port in self.WK_NONPRIV_PROTO_PORTS}
         nonpriv_ports.update({'other': int(not lowest_ports.issubset(self.WK_NONPRIV_PROTO_PORTS))})
         return nonpriv_ports
-
 
     def _get_priv_ports(self, rows, ip_proto, suffix):
 
@@ -236,27 +162,6 @@ class Host(Features):
 
         return self._host_rows(rows, priv_ports_present)
 
-
-    def tshark_priv_tcp_ports_in(self, rows):
-        rows = self._select_mac_direction(rows, output=False)
-        return self._get_priv_ports(rows, 'tcp', 'in')
-
-
-    def tshark_priv_tcp_ports_out(self, rows):
-        rows = self._select_mac_direction(rows, output=True)
-        return self._get_priv_ports(rows, 'tcp', 'out')
-
-
-    def tshark_priv_udp_ports_in(self, rows):
-        rows = self._select_mac_direction(rows, output=False)
-        return self._get_priv_ports(rows, 'udp', 'in')
-
-
-    def tshark_priv_udp_ports_out(self, rows):
-        rows = self._select_mac_direction(rows, output=True)
-        return self._get_priv_ports(rows, 'udp', 'out')
-
-
     def _get_nonpriv_ports(self, rows, ip_proto, suffix):
 
         def nonpriv_ports_present(host_rows):
@@ -265,27 +170,6 @@ class Host(Features):
                     for port, present in nonpriv_ports.items()}
 
         return self._host_rows(rows, nonpriv_ports_present)
-
-
-    def tshark_nonpriv_tcp_ports_in(self, rows):
-        rows = self._select_mac_direction(rows, output=False)
-        return self._get_nonpriv_ports(rows, 'tcp', 'in')
-
-
-    def tshark_nonpriv_tcp_ports_out(self, rows):
-        rows = self._select_mac_direction(rows, output=True)
-        return self._get_nonpriv_ports(rows, 'tcp', 'out')
-
-
-    def tshark_nonpriv_udp_ports_in(self, rows):
-        rows = self._select_mac_direction(rows, output=False)
-        return self._get_nonpriv_ports(rows, 'udp', 'in')
-
-
-    def tshark_nonpriv_udp_ports_out(self, rows):
-        rows = self._select_mac_direction(rows, output=True)
-        return self._get_nonpriv_ports(rows, 'udp', 'out')
-
 
     def _get_flags(self, rows, suffix, flags_field, decode_map):
         flags_counter = Counter()
@@ -301,12 +185,73 @@ class Host(Features):
             flags_field.replace('.', '_'), decoded_flag, suffix): val
                 for decoded_flag, val in flags_counter.items()}
 
-
     def _get_tcp_flags(self, rows, suffix):
         return self._get_flags(
             rows, suffix, 'tcp.flags',
             {0: 'fin', 1: 'syn', 2: 'rst', 3: 'psh', 4: 'ack', 5: 'urg', 6: 'ece', 7: 'cwr', 8: 'ns'})
 
+    def _get_ip_flags(self, rows, suffix):
+        return self._get_flags(
+            rows, suffix, 'ip.flags', {13: 'rb', 14: 'df', 15: 'mf'})
+
+    def _get_ip_dsfield(self, rows, suffix):
+        return self._get_flags(rows, suffix, 'ip.dsfield', {
+            0: 'ecn0', 1: 'ecn1', 2: 'dscp0', 3: 'dscp1', 4: 'dscp2',
+            5: 'dscp3', 6: 'dscp4', 7: 'dscp5'})
+
+    def tshark_last_protocols_array(self, rows):
+
+        def last_protocols_array(host_rows):
+            raw_protocols = set()
+            try:
+                raw_protocols.update({
+                    protocol for protocol in self._last_protocols(host_rows).split(':') if protocol})
+            except IndexError:
+                pass
+            raw_protocols -= set(['ethertype'])
+            protocols = {'protocol_%s' % protocol: int(protocol in raw_protocols) for protocol in self.WK_PROTOCOLS}
+            protocols.update({'other': int(not raw_protocols.issubset(self.WK_PROTOCOLS))})
+            return protocols
+
+        return self._host_rows(rows, last_protocols_array)
+
+    def tshark_ipv4(self, rows):
+        return self._host_rows(rows, lambda x: {'IPv4': int(4 in self._tshark_ipversions(x))})
+
+    def tshark_ipv6(self, rows):
+        return self._host_rows(rows, lambda x: {'IPv6': int(6 in self._tshark_ipversions(x))})
+
+    def tshark_priv_tcp_ports_in(self, rows):
+        rows = self._select_mac_direction(rows, output=False)
+        return self._get_priv_ports(rows, 'tcp', 'in')
+
+    def tshark_priv_tcp_ports_out(self, rows):
+        rows = self._select_mac_direction(rows, output=True)
+        return self._get_priv_ports(rows, 'tcp', 'out')
+
+    def tshark_priv_udp_ports_in(self, rows):
+        rows = self._select_mac_direction(rows, output=False)
+        return self._get_priv_ports(rows, 'udp', 'in')
+
+    def tshark_priv_udp_ports_out(self, rows):
+        rows = self._select_mac_direction(rows, output=True)
+        return self._get_priv_ports(rows, 'udp', 'out')
+
+    def tshark_nonpriv_tcp_ports_in(self, rows):
+        rows = self._select_mac_direction(rows, output=False)
+        return self._get_nonpriv_ports(rows, 'tcp', 'in')
+
+    def tshark_nonpriv_tcp_ports_out(self, rows):
+        rows = self._select_mac_direction(rows, output=True)
+        return self._get_nonpriv_ports(rows, 'tcp', 'out')
+
+    def tshark_nonpriv_udp_ports_in(self, rows):
+        rows = self._select_mac_direction(rows, output=False)
+        return self._get_nonpriv_ports(rows, 'udp', 'in')
+
+    def tshark_nonpriv_udp_ports_out(self, rows):
+        rows = self._select_mac_direction(rows, output=True)
+        return self._get_nonpriv_ports(rows, 'udp', 'out')
 
     def tshark_tcp_flags_in(self, rows):
 
@@ -316,7 +261,6 @@ class Host(Features):
 
         return self._host_rows(rows, tcp_flags_in)
 
-
     def tshark_tcp_flags_out(self, rows):
 
         def tcp_flags_out(host_rows):
@@ -324,12 +268,6 @@ class Host(Features):
             return self._get_tcp_flags(rows_filter, 'out')
 
         return self._host_rows(rows, tcp_flags_out)
-
-
-    def _get_ip_flags(self, rows, suffix):
-        return self._get_flags(
-            rows, suffix, 'ip.flags', {13: 'rb', 14: 'df', 15: 'mf'})
-
 
     def tshark_ip_flags_in(self, rows):
 
@@ -339,7 +277,6 @@ class Host(Features):
 
         return self._host_rows(rows, ip_flags_in)
 
-
     def tshark_ip_flags_out(self, rows):
 
         def ip_flags_out(host_rows):
@@ -347,13 +284,6 @@ class Host(Features):
             return self._get_ip_flags(rows_filter, 'out')
 
         return self._host_rows(rows, ip_flags_out)
-
-
-    def _get_ip_dsfield(self, rows, suffix):
-        return self._get_flags(rows, suffix, 'ip.dsfield', {
-            0: 'ecn0', 1: 'ecn1', 2: 'dscp0', 3: 'dscp1', 4: 'dscp2',
-            5: 'dscp3', 6: 'dscp4', 7: 'dscp5'})
-
 
     def tshark_ip_dsfield_in(self, rows):
 
@@ -363,7 +293,6 @@ class Host(Features):
 
         return self._host_rows(rows, ip_dsfield_in)
 
-
     def tshark_ip_dsfield_out(self, rows):
 
         def ip_dsfield_out(host_rows):
@@ -371,7 +300,6 @@ class Host(Features):
             return self._get_ip_dsfield(rows_filter, 'out')
 
         return self._host_rows(rows, ip_dsfield_out)
-
 
     def tshark_wk_ip_protos(self, rows):
 
@@ -389,7 +317,6 @@ class Host(Features):
 
         return self._host_rows(rows, protos)
 
-
     def tshark_vlan_id(self, rows):
 
         def first_vlan_id(host_rows):
@@ -402,7 +329,6 @@ class Host(Features):
 
         return self._host_rows(rows, first_vlan_id)
 
-
     def tshark_ipx(self, rows):
 
         def first_ipx(host_rows):
@@ -414,7 +340,6 @@ class Host(Features):
             return {'tshark_ipx': ipx}
 
         return self._host_rows(rows, first_ipx)
-
 
     def tshark_both_private_ip(self, rows):
 
@@ -432,7 +357,6 @@ class Host(Features):
 
         return self._host_rows(rows, first_not_both_private_ip)
 
-
     def tshark_ipv4_multicast(self, rows):
 
         def first_ipv4_multicast(host_rows):
@@ -446,7 +370,6 @@ class Host(Features):
             return {'tshark_ipv4_multicast': multicast}
 
         return self._host_rows(rows, first_ipv4_multicast)
-
 
     def tshark_non_ip(self, rows):
 
@@ -462,42 +385,32 @@ class Host(Features):
 
         return self._host_rows(rows, first_non_ip)
 
-
     def tshark_average_time_delta(self, rows):
         return self._calc_time_delta('average_time_delta', rows)
-
 
     def tshark_min_time_delta(self, rows):
         return self._calc_time_delta('min_time_delta', rows)
 
-
     def tshark_max_time_delta(self, rows):
         return self._calc_time_delta('max_time_delta', rows)
-
 
     def tshark_average_frame_len(self, rows):
         return self._calc_framelen('average_frame_len', rows)
 
-
     def tshark_min_frame_len(self, rows):
         return self._calc_framelen('min_frame_len', rows)
-
 
     def tshark_max_frame_len(self, rows):
         return self._calc_framelen('max_frame_len', rows)
 
-
     def tshark_median_frame_len(self, rows):
         return self._calc_framelen('median_frame_len', rows)
-
 
     def tshark_variance_frame_len(self, rows):
         return self._calc_framelen('variance_frame_len', rows)
 
-
     def tshark_25q_frame_len(self, rows):
         return self._calc_framelen('25q_frame_len', rows)
-
 
     def tshark_75q_frame_len(self, rows):
         return self._calc_framelen('75q_frame_len', rows)
@@ -507,142 +420,107 @@ class Host(Features):
     def tshark_min_frame_time_in(self, rows):
         return self._calc_tshark_field('min_frame_time_in', 'frame.time_epoch', rows)
 
-
     def tshark_min_frame_time_out(self, rows):
         return self._calc_tshark_field('min_frame_time_out', 'frame.time_epoch', rows)
-
 
     def tshark_max_frame_time_in(self, rows):
         return self._calc_tshark_field('max_frame_time_in', 'frame.time_epoch', rows)
 
-
     def tshark_max_frame_time_out(self, rows):
         return self._calc_tshark_field('max_frame_time_out', 'frame.time_epoch', rows)
-
 
     def tshark_count_frame_len_in(self, rows):
         return self._calc_framelen('count_frame_len_in', rows)
 
-
     def tshark_count_frame_len_out(self, rows):
         return self._calc_framelen('count_frame_len_out', rows)
-
 
     def tshark_total_frame_len_in(self, rows):
         return self._calc_framelen('total_frame_len_in', rows)
 
-
     def tshark_total_frame_len_out(self, rows):
         return self._calc_framelen('total_frame_len_out', rows)
-
 
     def tshark_average_frame_len_in(self, rows):
         return self._calc_framelen('average_frame_len_in', rows)
 
-
     def tshark_average_frame_len_out(self, rows):
         return self._calc_framelen('average_frame_len_out', rows)
-
 
     def tshark_25q_frame_len_in(self, rows):
         return self._calc_framelen('25q_frame_len_in', rows)
 
-
     def tshark_25q_frame_len_out(self, rows):
         return self._calc_framelen('25q_frame_len_out', rows)
-
 
     def tshark_75q_frame_len_in(self, rows):
         return self._calc_framelen('75q_frame_len_in', rows)
 
-
     def tshark_75q_frame_len_out(self, rows):
         return self._calc_framelen('75q_frame_len_out', rows)
-
 
     def tshark_median_frame_len_in(self, rows):
         return self._calc_framelen('median_frame_len_in', rows)
 
-
     def tshark_median_frame_len_out(self, rows):
         return self._calc_framelen('median_frame_len_out', rows)
-
 
     def tshark_variance_frame_len_in(self, rows):
         return self._calc_framelen('variance_frame_len_in', rows)
 
-
     def tshark_variance_frame_len_out(self, rows):
         return self._calc_framelen('variance_frame_len_out', rows)
-
 
     def tshark_max_frame_len_in(self, rows):
         return self._calc_framelen('max_frame_len_in', rows)
 
-
     def tshark_max_frame_len_out(self, rows):
         return self._calc_framelen('max_frame_len_out', rows)
-
 
     def tshark_min_frame_len_in(self, rows):
         return self._calc_framelen('min_frame_len_in', rows)
 
-
     def tshark_min_frame_len_out(self, rows):
         return self._calc_framelen('min_frame_len_out', rows)
-
 
     def tshark_min_time_delta_in(self, rows):
         return self._calc_time_delta('min_time_delta_in', rows)
 
-
     def tshark_min_time_delta_out(self, rows):
         return self._calc_time_delta('min_time_delta_out', rows)
-
 
     def tshark_25q_time_delta_in(self, rows):
         return self._calc_time_delta('25q_time_delta_in', rows)
 
-
     def tshark_25q_time_delta_out(self, rows):
         return self._calc_time_delta('25q_time_delta_out', rows)
-
 
     def tshark_median_time_delta_in(self, rows):
         return self._calc_time_delta('median_time_delta_in', rows)
 
-
     def tshark_median_time_delta_out(self, rows):
         return self._calc_time_delta('median_time_delta_out', rows)
-
 
     def tshark_average_time_delta_in(self, rows):
         return self._calc_time_delta('average_time_delta_in', rows)
 
-
     def tshark_average_time_delta_out(self, rows):
         return self._calc_time_delta('average_time_delta_out', rows)
-
 
     def tshark_75q_time_delta_in(self, rows):
         return self._calc_time_delta('75q_time_delta_in', rows)
 
-
     def tshark_75q_time_delta_out(self, rows):
         return self._calc_time_delta('75q_time_delta_out', rows)
-
 
     def tshark_max_time_delta_in(self, rows):
         return self._calc_time_delta('max_time_delta_in', rows)
 
-
     def tshark_max_time_delta_out(self, rows):
         return self._calc_time_delta('max_time_delta_out', rows)
 
-
     def tshark_variance_time_delta_in(self, rows):
         return self._calc_time_delta('variance_time_delta_in', rows)
-
 
     def tshark_variance_time_delta_out(self, rows):
         return self._calc_time_delta('variance_time_delta_out', rows)
